requirements.txt
matplotlib==3.1.1
networkx==2.3
dm-sonnet==1.34
numpy==1.16.4
tensorflow==1.14
six==1.12
tensorflow-probability==0.7.0
tensorflow_estimator==1.14.0
keras==2.3.0
jupyter